% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{field_maps_qry}
\alias{field_maps_qry}
\title{Generate SQL query for administrative boundary data}
\usage{
field_maps_qry(iso3, level, con, dataset = "humanitarian", geom = "polygons")
}
\arguments{
\item{iso3}{A character string containing the ISO3 country code}

\item{level}{An integer specifying the administrative level (1, 2, 3 or 4)}

\item{con}{A DBI duckdb connection object for SQL safety}

\item{dataset}{A character string specifying the dataset type, either
"humanitarian" (default) or "open"}

\item{geom}{A character string specifying the geometry type, either
"polygons" (default), "lines" or "points"}
}
\value{
A character string containing the SQL query
}
\description{
Creates a SQL query string to retrieve administrative boundary data from
Field Maps remote parquet files based on specified parameters.
}
\keyword{internal}
