% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldmaps.R
\name{get_all_adm_levels}
\alias{get_all_adm_levels}
\title{Get All Available Administrative Levels for a Country}
\usage{
get_all_adm_levels(
  country,
  dataset = "humanitarian",
  geom = "polygons",
  con = NULL
)
}
\arguments{
\item{country}{Character. Country name or ISO3 code to download boundaries for.}

\item{dataset}{Character. Dataset to use, either "humanitarian" or "open".
Default is "humanitarian". See https://fieldmaps.io/data/ for details.}

\item{geom}{Character. Geometry type to retrieve: "polygons", "lines", or "points".
Default is "polygons".}

\item{con}{DuckDB connection object. If NULL, a temporary connection will be created.}
}
\value{
A named list of sf objects, with names like "ADM1", "ADM2", etc.
Each sf object contains administrative boundaries for that level with CRS 4326 (WGS84).
}
\description{
Downloads all available administrative boundary levels for a country from
the Field Maps dataset. This function automatically determines the maximum
administrative level available and retrieves all levels from 1 to the maximum.
}
\examples{
\dontrun{
# Get all administrative levels for Somalia
somalia_all <- get_all_adm_levels("Somalia")

# Access individual levels
somalia_adm1 <- somalia_all$ADM1
somalia_adm2 <- somalia_all$ADM2

# Get all levels from open dataset
boundaries <- get_all_adm_levels("ETH", dataset = "open")
}

}
